move navigators to separate file
